backend/Idempotency-gateway/requirements.txt
Django==5.2.13
djangorestframework==3.17.1